phoenix: show toast for duration of focusTitle() call
I think that this does nothing. It seems like the toast never actually
shows because the Phoenix is busy processing other stuff, trying to
focus on the specified window. Then Phoenix finishes with the process
and is now able to deal with the toast that needs to be displayed,
except that now, it's time to close the toast, because the process is
finished.