vim: Use vim-javascript plugin conditionally
Vim-javascript is much slower than vim's built-in js syntax. I don't
want to give it up though, because vim's built-in js syntax is rather
dumb. So I'm giving myself the option to quickly turn it on and off so I
can edit huge js files without hangups but still be able to use advanced
syntax highlighting most of the time.